updated project

+++ src/main/java/com/Satya/SpringBoot/Projects/Ecommerce/Rest/API/Controler/ProductControler.java
@@ -1,12 +1,10 @@
 package com.Satya.SpringBoot.Projects.Ecommerce.Rest.API.Controler;
 
-import com.Satya.SpringBoot.Projects.Ecommerce.Rest.API.Model.Products;
+import com.Satya.SpringBoot.Projects.Ecommerce.Rest.API.Model.Product;
 import com.Satya.SpringBoot.Projects.Ecommerce.Rest.API.Service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -22,13 +20,13 @@ public class ProductControler {
     private ProductService productService;
 
     @GetMapping("products")
-    public List<Products> getAllProducts() {
+    public List<Product> getAllProducts() {
         return productService.showAll();
     }
 
     @GetMapping("/product/{productId}/image")
     public ResponseEntity<byte[]> getProductImage(@PathVariable int productId) {
-        Products product = productService.getProduct(productId);
+        Product product = productService.getProduct(productId);
         if (product.getId() > 0) {
             return new ResponseEntity<>(product.getImageData(), HttpStatus.OK);
         } else {
@@ -38,10 +36,10 @@ public class ProductControler {
 
 
     @PostMapping("product")
-    public ResponseEntity<?> addProduct(@RequestPart Products product, @RequestPart MultipartFile imageFile) {
-        Products saveProduct= null;
+    public ResponseEntity<?> addProduct(@RequestPart Product product, @RequestPart MultipartFile imageFile) {
+        Product saveProduct= null;
         try {
-            saveProduct = productService.addProduct(product,imageFile);
+            saveProduct = productService.saveOrUpdate(product,imageFile);
             return new ResponseEntity<>(saveProduct,HttpStatus.CREATED);
 
         } catch (IOException e) {
@@ -50,19 +48,28 @@ public class ProductControler {
 
     }
 
-    @DeleteMapping("delete/{id}")
-    public String DeleteProduct(@PathVariable int id) {
-        productService.deleteProduct(id);
-        return "Deleted";
+    @PutMapping("/product/{id}")
+    public ResponseEntity<String> updateProduct(@PathVariable int id, @RequestPart Product product, @RequestPart MultipartFile imageFile) {
+
+        Product updatedpPoduct =null;
+        try{
+           updatedpPoduct= productService.saveOrUpdate(product,imageFile);
+           return new ResponseEntity<>("Updated",HttpStatus.OK);
+        } catch (Exception e){
+            return new ResponseEntity<>(e.getMessage(),HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
-    @PutMapping("/product/update/${id}")
-    public ResponseEntity<?> updateProduct(@PathVariable int id) {
+    @DeleteMapping("/product/{id}")
+    public ResponseEntity<String> deleteProduct(@PathVariable int id) {
+        Product product = productService.getProduct(id);
+        if (product != null) {
+            productService.deleteProduct(id);
+            return new ResponseEntity<>("Deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND);
+        }}
 
-        Products product =productService.getProduct(id);
-
-        return new ResponseEntity<>(productService.updateProduct(product));
-    }
 
     @GetMapping("load")
     public ResponseEntity<String> load() {           //ResponseEntity --> Status Code
@@ -71,14 +78,21 @@ public class ProductControler {
     }
 
     @GetMapping("product/{id}")
-    public ResponseEntity<Products> getProduct(@PathVariable int id) {
-        Products product = productService.getProduct(id);
+    public ResponseEntity<Product> getProduct(@PathVariable int id) {
+        Product product = productService.getProduct(id);
 
         if(product.getId() >0){
             return new ResponseEntity<>(product,HttpStatus.OK);
         }
         else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
+    }
+
+    @GetMapping("/product/search")
+    public ResponseEntity<List<Product>> searchProducts(@RequestParam String keyword) {
+        List<Product> products = productService.searchProducts(keyword);
+        System.out.println("searching with :" + keyword);
+        return new ResponseEntity<>(products, HttpStatus.OK);
     }
